requirements.txt
selenium==3.141.0
pandas==1.3.0
google-cloud-bigquery==2.26.0
pyarrow==12.0.1

